+++ lib/quickeebooks/windows/model/intuit_type.rb
@@ -7,6 +7,6 @@ module Quickeebooks
         include Logging
 
       end 
-    end    
+    end
   end
 end